W-10619644: Split ConnectionProviderModelValidator (#11500)
